requirements.txt
google-generativeai
marko
neo4j
Flask
flask_cors
PyPDF2
python-docx